refactor: Operator∙StringCalculator 수정, Converter 추가
Operator 내부 BiFunction 추가 및 if문 제거
calculate() 리턴 타입을 double로 수정
StringCalculator에서 사용하는 Converter 추가

+++ src/main/java/caculator/CalculatorApplication.java
@@ -6,7 +6,7 @@ public class CalculatorApplication {
         Input input = new Input();
         String inputValue = input.input("문자열을 입력하세요: ");
         StringCalculator stringCalculator = new StringCalculator();
-        int calculate = stringCalculator.calculate(inputValue);
+        double calculate = stringCalculator.calculate(inputValue);
         System.out.println("계산 결과: " + calculate);
     }
 }
